+++ base_exts/src/alist.ml
@@ -31,15 +31,18 @@ module Extensions = struct
   include (M : module type of M with type ('k, 'v) t := ('k, 'v) t)
 
   include Bi_mappable.Extend2 (M)
+
+  let compose_match (type a b c) (k : a) (v : b) (k' : b) (v' : c)
+      ~(equal : b -> b -> bool) : (a * c) option =
+    Option.some_if (equal v k') (k, v')
+
+  let compose_one (type a b c) (bc : (b, c) t) ((k, v) : a * b)
+      ~(equal : b -> b -> bool) : (a * c) list =
+    List.filter_map bc ~f:(fun (k', v') -> compose_match k v k' v' ~equal)
+
+  let compose (type a b c) (ab : (a, b) t) (bc : (b, c) t)
+      ~(equal : b -> b -> bool) : (a, c) t =
+    List.concat_map ab ~f:(compose_one bc ~equal)
 end
 
 include Extensions
-
-let%expect_test "bi_map example" =
-  let sample = [("foo", 27); ("bar", 53); ("baz", 99)] in
-  let sample' = bi_map sample ~left:String.capitalize ~right:Int.neg in
-  List.iter ~f:(fun (k, v) -> Stdio.printf "%s -> %d\n" k v) sample' ;
-  [%expect {|
-    Foo -> -27
-    Bar -> -53
-    Baz -> -99 |}]